requirements.txt
numpy==1.26.0
opencv_python==4.10.0.82
Pillow==10.3.0
gui==0.0.1
utils==1.0.2
